Move tests to /tests directory and use unittest instead of pytest
The motivation for this commit was mainly the usage of unittest. Pytest
is a third-party package which requires that a user of this project have
Pytest installed. Unittest is part of Python's standard framework.
Additionally, Unittest removes redundant code seen in the Pytest tests.
For example, recreating a list for each test in binary search was moved
to the `setUp` method. Finally, the tests can be run with 1 command
executed from the directory `/tests`:

`python -m unittest discover`

Python will search for and automatically run all files prefixed with
"test".